Always include <gmp.h> before including relic headers (#21)
The relic headers internally include <gmp.h> as well. At the same time,
inclusion of relic headers is wrapped into a namespace, which makes the
gmp headers be part of that namespace as well. This leads to conflicting
definitions of "std::ostream& operator<< (std::ostream &, mpz_srcptr)" and
friends on Debian stretch (if gmp from /usr is used).