Fix invalid access to data in rvalues when converting py::bytes
Conversion like "std::string(b)" result in a temporary r value with a
short/undefined lifetime. Accessing data of it via pointers into internal
data gives undefined behavior.

The fix is to use a temporary variable which has a well defined lifetime.

The issue that is fixed here is that even when "from_xxx" methods were
called with the same data, the resulting keys were different each time.
I assume this was not noticed before as it only appear to happen with
some compiler versions.